refactor(json): split the scalar emitters out of stringify.rs (2000-line cap)
This PR pushed stringify.rs to 2041 lines and the file-size lint gate rejected
it. Moved the number formatter, the string escaper and the BigInt path into
stringify_scalars.rs — a coherent group with no dependency on the traversal
around them — and re-exported them, so no call site changes.

  stringify.rs        2041 -> 1787
  stringify_scalars.rs        269

Pure code move; no behaviour change. Mirrors the existing stringify_api.rs /
stringify_buffer.rs split in the same directory.
